refactor: share compiler file-system path helpers
